requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
yt-dlp==2024.4.9
aiohttp==3.9.5
aiofiles==23.2.1
python-dotenv==1.0.1
pillow==10.3.0
psutil==5.9.8
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
